create EventBusModule to wrap using Nests CQRS backed by RabbitMQ
