--- src/domain/community/community/events/virtual-contributor-added.event.ts
@@ -1,8 +1,0 @@
-import { IEvent } from '@nestjs/cqrs';
-
-export class VirtualContributorAdded implements IEvent {
-  constructor(
-    public readonly spaceId: string,
-    public readonly contributorId: string
-  ) {}
-}
