+++ src/services/event-bus/commands/index.ts
@@ -1,0 +1,1 @@
+export { IngestSpace, SpaceIngestionPurpose } from './ingest.space.command';
